+++ src/dlls/objects/471_DIM2SnowBal/DIM2SnowBal.c
@@ -7,6 +7,8 @@
 #include "dlls/objects/471_DIM2SnowBal.h"
 #include "dlls/objects/472_DIM2PathGen.h"
 
+#include "main/audio/sfx_keep_alive_api.h"
+#include "main/audio/sfx_play_api.h"
 #include "main/audio/sfx_trigger_ids.h"
 #include "main/curve.h"
 #include "main/dll/partfx_interface.h"
@@ -16,7 +18,9 @@
 #include "main/object_render.h"
 #include "main/track_bbox_api.h"
 #include "main/track_dolphin_api.h"
+#include "main/vecmath.h"
 #include "sys/objects.h"
+#include "sys/objects/lifecycle.h"
 
 #define DIM2_SNOWBALL_SHARPCLAW_SEQUENCE_ID 214
 #define DIM2_SNOWBALL_IMPACT_PARTFX_ID      518
@@ -119,7 +123,7 @@ void dim2snowball_update(GameObject* obj) {
                 if (sharpClaw != NULL) {
                     (*(void (**)(GameObject*))(*(int*)sharpClaw->anim.dll + 0x20))(sharpClaw);
                 }
-                Sfx_PlayFromObject((int)obj, SFXTRIG_en_nlite1_c);
+                Sfx_PlayFromObject(obj, SFXTRIG_en_nlite1_c);
             }
 
             particleParams.posX = obj->anim.localPosX;
