+++ src/dlls/objects/489_SB_Propelle/SB_Propelle.c
@@ -12,10 +12,15 @@
 #include "dlls/objects/489_SB_Propelle.h"
 
 #include "dlls/objects/488_SB_Galleon.h"
+#include "main/audio/sfx_keep_alive_api.h"
+#include "main/audio/sfx_play_api.h"
 #include "main/dll/partfx_interface.h"
 #include "main/frame_timing.h"
+#include "main/obj_path.h"
 #include "main/object_render.h"
 #include "main/objfx.h"
+#include "main/objhits.h"
+#include "main/vecmath.h"
 #include "sys/objects.h"
 
 /* anim.romDefNo tag identifying a live propeller (vs. a placeholder stand-in) */
@@ -122,7 +127,7 @@ void SB_Propeller_update(GameObject* obj) {
             (obj->userData1 = 0x14, obj->anim.parent != NULL) && (galleonStage == 2 || galleonStage == 5) &&
             obj->anim.romDefNo == SB_PROPELLER_SEQ_ID) {
             Obj_SetModelColorFadeRecursive(obj, 0xf, 200, 0, 0, 1);
-            Sfx_PlayFromObject((int)obj, SB_PROPELLER_SFX_HIT);
+            Sfx_PlayFromObject(obj, SB_PROPELLER_SFX_HIT);
             state->health -= 1;
             if (state->health <= 0) {
                 state->health = 0;
@@ -130,7 +135,7 @@ void SB_Propeller_update(GameObject* obj) {
                 ObjHits_DisableObject(obj);
                 obj->anim.flags = obj->anim.flags | OBJANIM_FLAG_HIDDEN;
                 spawnExplosion(obj, 100.0f, 1, 1, 1, 0, 1, 1, 0);
-                Sfx_PlayFromObject((int)obj, SB_PROPELLER_SFX_DESTROYED);
+                Sfx_PlayFromObject(obj, SB_PROPELLER_SFX_DESTROYED);
             }
         }
         if (obj->userData1 == 0) {
